Added Get/Set functions.
Changed the way the database reacts with the Main script.
Added Get/Set functions and Removed the Entity.SetData/GetData
Added List for Players and Player stats and attached them via SQL ID

On Login, Player is created and Given the ID from the database via client.SetData
The Client when changing values for, is called by his/her ID with a loop through the Player list, At that point the data is edited when the loop finds the specific Player Instance with that given ID,

Whenever there are functions using the players data The main script must be called.

-Mars